requirements.txt
apache-airflow==2.7.3
boto3
psycopg2-binary
python-dotenv